+++ src/app/layout.tsx
@@ -4,6 +4,14 @@ import { Inter } from 'next/font/google'
 
 const inter = Inter({ subsets: ['latin'] })
 
+const mainAreaStyle:React.CSSProperties = {
+  display: 'flex',
+  height: '100vh',
+  backgroundColor: 'rgb(235, 228, 209)',
+  marginTop: '72px',
+  alignItems: 'center',
+  flexDirection: 'column'
+}
 export const metadata: Metadata = {
   title: 'Create Next App',
   description: 'Generated by create next app',
