+++ src/app/login/page.tsx
@@ -1,12 +1,12 @@
 import Login from "@/app/components/Login";
-import MainArea from "@/app/components/MainArea";
+//import MainArea from "@/app/components/MainArea";
 import TopBar from "@/app/components/TopBar";
 
 const Page = () => {
     return (
-        <div>
+        <div className="mainArea">
             <TopBar/>
-            <MainArea/>
+            <Login/>
         </div>
     )
 }
